+++ Framework/PWLocation.framework/Versions/A/Headers/PWSLLocationManager.h
@@ -49,8 +49,17 @@ The map identifier obtained from Phunware. (read-only)
  @discussion If the device is not within 5 kilometers of the venue location, location updates will fail.
  @return The location manager object.
  */
+- (instancetype)initWithMapIdentifier:(NSString *)mapID customerIdentifier:(NSString *)customerID location:(CLLocationCoordinate2D)venueLocation __attribute__((deprecated));
 
-- (instancetype)initWithMapIdentifier:(NSString *)mapID customerIdentifier:(NSString *)customerID location:(CLLocationCoordinate2D)venueLocation;
+
+/**
+ Initializes the location manager with the specified venue GUID..
+ @param mapID - Obtained from Phunware.
+ @param customerID - Obtained from Phunware.
+ @discussion If the device is not within 5 kilometers of the venue location, location updates will fail.
+ @return The location manager object.
+ */
+- (instancetype)initWithMapIdentifier:(NSString *)mapID customerIdentifier:(NSString *)customerID NS_DESIGNATED_INITIALIZER;
 
 - (BOOL)locationServicesAvailable;
 - (void)startUpdatingLocation;